requirements.txt
transformers>=4.30.0
huggingface-hub>=0.16.0
torch>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
tqdm>=4.65.0
pyyaml>=6.0
